compute/wps/tests/requirements.txt
mock
nose
django-nose
coverage
